+++ src/model/types.rs
@@ -2,7 +2,7 @@
 //!
 //! `model::types` defines custom type structs with trait implementations corresponding to Ruddr API custom types.
 use regex::Regex;
-use serde::{Deserialize, Deserializer};
+use serde::{Deserializer, Serialize};
 use std::fmt;
 
 /// Custom type for Ruddr Date type in YYYY-MM-DD format.
@@ -10,7 +10,7 @@ use std::fmt;
 /// ```ignore
 /// Date::try_from("2028-12-31")
 /// ```
-#[derive(PartialEq, Eq, Debug, Clone)]
+#[derive(PartialEq, Eq, Debug, Clone, Serialize)]
 // public access to the type should exist, but not to the implicit constructor as users are expected to instantiate through type converters each containing an invocation to the explicit constructor
 pub struct Date(pub(super) String);
 
@@ -79,7 +79,7 @@ impl fmt::Display for Date {
 /// Timestamp::try_from("1234-56-78T12:34:56.789Z")
 /// ```
 /// This type is currently unused in input parameters to interface functions, but is still public for potential unforeseen usage.
-#[derive(PartialEq, Eq, Debug, Clone)]
+#[derive(PartialEq, Eq, Debug, Clone, Serialize)]
 // public access to the type should exist, but not to the implicit constructor as users are expected to instantiate through type converters each containing an invocation to the explicit constructor
 pub struct Timestamp(pub(super) String);
 
@@ -147,7 +147,7 @@ impl fmt::Display for Timestamp {
 /// ```ignore
 /// Time::try_from("12:34")
 /// ```
-#[derive(PartialEq, Eq, Debug, Clone)]
+#[derive(PartialEq, Eq, Debug, Clone, Serialize)]
 // public access to the type should exist, but not to the implicit constructor as users are expected to instantiate through type converters each containing an invocation to the explicit constructor
 pub struct Time(pub(super) String);
 
@@ -214,7 +214,7 @@ impl fmt::Display for Time {
 /// ```ignore
 /// UUID::try_from("4c8d3f42-6efd-4a7e-85ca-d43164db0ab2")
 /// ```
-#[derive(PartialEq, Eq, Deserialize, Debug, Clone)]
+#[derive(PartialEq, Eq, Debug, Clone, Serialize)]
 // public access to the type should exist, but not to the implicit constructor as users are expected to instantiate through type converters each containing an invocation to the explicit constructor
 pub struct UUID(pub(super) String);
 
@@ -261,13 +261,13 @@ impl<'uuid> From<&'uuid UUID> for &'uuid str {
     }
 }
 
-/*impl<'de> serde::Deserialize<'de> for UUID {
+impl<'de> serde::Deserialize<'de> for UUID {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        Ok(UUID::new(String::deserialize(deserializer)?))
+        UUID::new(String::deserialize(deserializer)?).map_err(serde::de::Error::custom)
     }
 }
 
-impl serde::Serialize for UUID {
+/*impl serde::Serialize for UUID {
     fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
         serializer.serialize_str(&self.0)
     }
@@ -284,7 +284,7 @@ impl fmt::Display for UUID {
 /// ```ignore
 /// Slug::try_from("vendor-portal")
 /// ```
-#[derive(PartialEq, Eq, Debug, Clone)]
+#[derive(PartialEq, Eq, Debug, Clone, Serialize)]
 // public access to the type should exist, but not to the implicit constructor as users are expected to instantiate through type converters each containing an invocation to the explicit constructor
 pub struct Slug(pub(super) String);